Submission json file for test set batch 4
File containing all predictions we made with the version of the model that we had by 9th of May, when we attempted to participate in the competition with our model. Unfortunately, our file was not accepted because of a small mismatch in the json file format used which led to our submission file to be rejected by their algorithm. However, the predicted answer showcase that our models make relevant predictions with a low rate of faulty answers.